refactor: upgrade ArcGISProToArcGISEarth project to .Net 6
