+++ src/ArcGISEarth.AutoAPI.Samples.sln
@@ -1,38 +1,38 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30323.103
+# Visual Studio Version 17
+VisualStudioVersion = 17.7.33808.371
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ArcGISEarth.AutoAPI.Examples", "ArcGISEarth.AutoAPI.Examples\ArcGISEarth.AutoAPI.Examples.csproj", "{E6B66EF3-7476-4B46-8B15-468E77065F46}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ArcGISProToArcGISEarth", "ArcGISProToArcGISEarth\ArcGISProToArcGISEarth.csproj", "{94776263-9F12-4574-927E-C3C2DE42BCAF}"
-EndProject
 Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "ArcGISEarth.AutoAPI.Shared", "ArcGISEarth.AutoAPI.Shared\ArcGISEarth.AutoAPI.Shared.shproj", "{5159E20A-F3B9-4D45-B27E-22D3CABCC8AB}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ArcGISEarth.AutoAPI.Examples.NetCore", "ArcGISEarth.AutoAPI.Examples\ArcGISEarth.AutoAPI.Examples.NetCore.csproj", "{7949AAE8-EC91-4280-86AE-42FA496A7EF9}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ArcGISProToArcGISEarth", "ArcGISProToArcGISEarth\ArcGISProToArcGISEarth.csproj", "{122173C8-AB14-4A2E-97FC-C0C704092EDB}"
+EndProject
 Global
-	GlobalSection(SharedMSBuildProjectFiles) = preSolution
-		ArcGISEarth.AutoAPI.Shared\ArcGISEarth.AutoAPI.Shared.projitems*{5159e20a-f3b9-4d45-b27e-22d3cabcc8ab}*SharedItemsImports = 13
-		ArcGISEarth.AutoAPI.Shared\ArcGISEarth.AutoAPI.Shared.projitems*{94776263-9f12-4574-927e-c3c2de42bcaf}*SharedItemsImports = 4
-		ArcGISEarth.AutoAPI.Shared\ArcGISEarth.AutoAPI.Shared.projitems*{e6b66ef3-7476-4b46-8b15-468e77065f46}*SharedItemsImports = 4
-	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E6B66EF3-7476-4B46-8B15-468E77065F46}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E6B66EF3-7476-4B46-8B15-468E77065F46}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E6B66EF3-7476-4B46-8B15-468E77065F46}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E6B66EF3-7476-4B46-8B15-468E77065F46}.Release|Any CPU.Build.0 = Release|Any CPU
-		{94776263-9F12-4574-927E-C3C2DE42BCAF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{94776263-9F12-4574-927E-C3C2DE42BCAF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{94776263-9F12-4574-927E-C3C2DE42BCAF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{94776263-9F12-4574-927E-C3C2DE42BCAF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7949AAE8-EC91-4280-86AE-42FA496A7EF9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7949AAE8-EC91-4280-86AE-42FA496A7EF9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7949AAE8-EC91-4280-86AE-42FA496A7EF9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7949AAE8-EC91-4280-86AE-42FA496A7EF9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{122173C8-AB14-4A2E-97FC-C0C704092EDB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{122173C8-AB14-4A2E-97FC-C0C704092EDB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{122173C8-AB14-4A2E-97FC-C0C704092EDB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{122173C8-AB14-4A2E-97FC-C0C704092EDB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {70DAF370-3DE8-4599-BDFC-51EE548C34E7}
+	EndGlobalSection
+	GlobalSection(SharedMSBuildProjectFiles) = preSolution
+		ArcGISEarth.AutoAPI.Shared\ArcGISEarth.AutoAPI.Shared.projitems*{122173c8-ab14-4a2e-97fc-c0c704092edb}*SharedItemsImports = 5
+		ArcGISEarth.AutoAPI.Shared\ArcGISEarth.AutoAPI.Shared.projitems*{5159e20a-f3b9-4d45-b27e-22d3cabcc8ab}*SharedItemsImports = 13
+		ArcGISEarth.AutoAPI.Shared\ArcGISEarth.AutoAPI.Shared.projitems*{7949aae8-ec91-4280-86ae-42fa496a7ef9}*SharedItemsImports = 5
 	EndGlobalSection
 EndGlobal
