+++ src/ArcGISProToArcGISEarth/ArcGISEarthModule.cs
@@ -1,36 +1,47 @@
-﻿// Copyright 2020 Esri
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
+﻿using ArcGIS.Core.CIM;
+using ArcGIS.Core.Data;
+using ArcGIS.Core.Geometry;
+using ArcGIS.Desktop.Catalog;
+using ArcGIS.Desktop.Core;
+using ArcGIS.Desktop.Editing;
+using ArcGIS.Desktop.Extensions;
 using ArcGIS.Desktop.Framework;
 using ArcGIS.Desktop.Framework.Contracts;
+using ArcGIS.Desktop.Framework.Dialogs;
+using ArcGIS.Desktop.Framework.Threading.Tasks;
+using ArcGIS.Desktop.Layouts;
+using ArcGIS.Desktop.Mapping;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Input;
 
-namespace ToArcGISEarth
+namespace ArcGISProToArcGISEarth
 {
     internal class ArcGISEarthModule : Module
     {
         private static ArcGISEarthModule _this = null;
 
-        public static ArcGISEarthModule Current
-        {
-            get
-            {
-                return _this ?? (_this = (ArcGISEarthModule)FrameworkApplication.FindModule("ArcGISProToArcGISEarth"));
-            }
-        }
+        /// <summary>
+        /// Retrieve the singleton instance to this module here
+        /// </summary>
+        public static ArcGISEarthModule Current => _this ??= (ArcGISEarthModule)FrameworkApplication.FindModule("ArcGISEarthModule");
 
+        #region Overrides
+        /// <summary>
+        /// Called by Framework when ArcGIS Pro is closing
+        /// </summary>
+        /// <returns>False to prevent Pro from closing, otherwise True</returns>
         protected override bool CanUnload()
         {
+            //TODO - add your business logic
+            //return false to ~cancel~ Application close
             return true;
         }
+
+        #endregion Overrides
+
     }
 }
