+++ src/ArcGISProToArcGISEarth/Config.daml
@@ -1,55 +1,57 @@
-﻿<ArcGIS defaultAssembly="ToArcGISEarth.dll" defaultNamespace="ToArcGISEarth" xmlns="http://schemas.esri.com/DADF/Registry" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://schemas.esri.com/DADF/Registry file:///C:/Program%20Files/ArcGIS/Pro/bin/ArcGIS.Desktop.Framework.xsd">
-  <AddInInfo id="{94776263-9f12-4574-927e-c3c2de42bcaf}" version="1.2" desktopVersion="2.5 or later">
-    <Name>ToArcGISEarth</Name>
+﻿<ArcGIS defaultAssembly="ArcGISProToArcGISEarth.dll" defaultNamespace="ArcGISProToArcGISEarth" xmlns="http://schemas.esri.com/DADF/Registry" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://schemas.esri.com/DADF/Registry file:///C:/Program%20Files/ArcGIS/Pro/bin/ArcGIS.Desktop.Framework.xsd">
+  <AddInInfo id="{d754c583-5aa9-4b29-8b9e-c04afdb2d2c3}" version="1.0" desktopVersion="3.2.48173">
+    <Name>ArcGISProToArcGISEarth</Name>
     <Description>ArcGIS Pro to ArcGIS Earth</Description>
     <Image>Images\ArcGISEarth32.png</Image>
     <Author>ArcGIS Earth Tutorials</Author>
     <Company>Esri, http://www.esri.com</Company>
-    <Date>10/01/2020 11:00:00 AM, 2020</Date>
+    <Date>06/26/2020 11:00:00 AM, 2023</Date>
     <Subject>Framework</Subject>
   </AddInInfo>
   <modules>
-    <insertModule id="ArcGISProToArcGISEarth" className="ArcGISEarthModule" autoLoad="false" caption="ArcGISEarthModule">
+    <insertModule id="ArcGISProToArcGISEarth" className="ArcGISEarthModule" autoLoad="true" caption="ArcGISEarthModule">
       <groups>
-        <group id="ToArcGISEarth_Group1" caption="To ArcGIS Earth" appearsOnAddInTab="true" keytip="TE">          
-          <menu refID="ToArcGISEarth_SynchronizationMenu" size="middle"/>
+        <group id="ToArcGISEarth_Group1" caption="To ArcGIS Earth" appearsOnAddInTab="true" keytip="TE">
+          <menu refID="ToArcGISEarth_SynchronizationMenu" size="middle" />
           <button refID="ToArcGISEarth_ClearWorkspaceButton" size="middle" />
           <button refID="ToArcGISEarth_SaveImageButton" size="middle" />
           <button refID="ToArcGISEarth_HelpButton" size="middle" />
+          <button refID="ArcGISProToArcGISEarth_Button1" size="large" />
+          <button refID="ArcGISProToArcGISEarth_Button1" size="large" />
         </group>
       </groups>
       <menus>
-        <menu id="ToArcGISEarth_SynchronizationMenu" caption="Synchronization" smallImage="Images\Synchronization16.png" largeImage="Images\Synchronization32.png" keytip="SZ" >
+        <menu id="ToArcGISEarth_SynchronizationMenu" caption="Synchronization" smallImage="pack://application:,,,/ArcGISProToArcGISEarth;component/Images/Synchronization16.png" largeImage="pack://application:,,,/ArcGISProToArcGISEarth;component/Images/Synchronization32.png" keytip="SZ">
           <tooltip>Synchronization operations between ArcGIS Pro and ArcGIS Earth.</tooltip>
           <button refID="ToArcGISEarth_SetCameraSyncCheckBox" />
           <button refID="ToArcGISEarth_AddLayerSyncCheckBox" />
           <button refID="ToArcGISEarth_RemoveLayerSyncCheckBox" />
         </menu>
       </menus>
-      <controls>       
-        <button id="ToArcGISEarth_SetCameraSyncCheckBox" caption="Camera view" className="SetCameraSyncCheckBox" loadOnClick="false"  keytip="SC">
+      <controls>
+        <button id="ToArcGISEarth_SetCameraSyncCheckBox" caption="Camera view" className="SetCameraSyncCheckBox" loadOnClick="false" keytip="SC">
           <tooltip>Synchronize camera view with ArcGIS Earth.</tooltip>
         </button>
-        <button id="ToArcGISEarth_AddLayerSyncCheckBox" caption="Add layer" className="AddLayerSyncCheckBox" loadOnClick="false"  keytip="AL">
+        <button id="ToArcGISEarth_AddLayerSyncCheckBox" caption="Add layer" className="AddLayerSyncCheckBox" loadOnClick="false" keytip="AL">
           <tooltip>Synchronize the operation when adding layer.</tooltip>
         </button>
-        <button id="ToArcGISEarth_RemoveLayerSyncCheckBox" caption="Remove layer" className="RemoveLayerSyncCheckBox" loadOnClick="false"  keytip="RL">
+        <button id="ToArcGISEarth_RemoveLayerSyncCheckBox" caption="Remove layer" className="RemoveLayerSyncCheckBox" loadOnClick="false" keytip="RL">
           <tooltip>Synchronize the operation when removing layer.</tooltip>
         </button>
-        <button id="ToArcGISEarth_ClearWorkspaceButton" caption="Clear workspace" className="ClearWorkspaceButton" loadOnClick="false" smallImage="Images\ClearWorkspace16.png" largeImage="Images\ClearWorkspace32.png" keytip="CW">
+        <button id="ToArcGISEarth_ClearWorkspaceButton" caption="Clear workspace" className="ClearWorkspaceButton" loadOnClick="false" smallImage="pack://application:,,,/ArcGISProToArcGISEarth;component/Images/ClearWorkspace16.png" largeImage="pack://application:,,,/ArcGISProToArcGISEarth;component/Images/ClearWorkspace32.png" keytip="CW">
           <tooltip>Clear workspace in ArcGIS Earth.</tooltip>
         </button>
-        <button id="ToArcGISEarth_SaveImageButton" caption="Save image" className="SaveImageButton" loadOnClick="false" smallImage="Images\SaveImage16.png" largeImage="Images\SaveImage32.png" keytip="SI">
+        <button id="ToArcGISEarth_SaveImageButton" caption="Save image" className="SaveImageButton" loadOnClick="false" smallImage="pack://application:,,,/ArcGISProToArcGISEarth;component/Images/SaveImage16.png" largeImage="pack://application:,,,/ArcGISProToArcGISEarth;component/Images/SaveImage32.png" keytip="SI">
           <tooltip>Save current view as image in ArcGIS Earth.</tooltip>
         </button>
-        <button id="ToArcGISEarth_HelpButton" caption="Help" className="HelpButton" loadOnClick="false" smallImage="Images\Help16.png" largeImage="Images\Help32.png" keytip="HE">
+        <button id="ToArcGISEarth_HelpButton" caption="Help" className="HelpButton" loadOnClick="false" smallImage="pack://application:,,,/ArcGISProToArcGISEarth;component/Images/Help16.png" largeImage="pack://application:,,,/ArcGISProToArcGISEarth;component/Images/Help32.png" keytip="HE">
           <tooltip>See ArcGIS Earth Automation API help document for more details.</tooltip>
+        </button>
+        <button id="ArcGISProToArcGISEarth_Button1" caption="Button 1" className="Button1" loadOnClick="true" smallImage="pack://application:,,,/ArcGIS.Desktop.Resources;component/Images/GenericButtonBlue16.png" largeImage="pack://application:,,,/ArcGIS.Desktop.Resources;component/Images/GenericButtonBlue32.png">
+          <tooltip heading="Tooltip Heading">Tooltip text<disabledText /></tooltip>
         </button>
       </controls>
       <panes></panes>
     </insertModule>
   </modules>
-  <dropHandlers>
-    <insertHandler id="ToArcGISEarth_DropHandler1" className="DropHandler1" dataTypes="*" />
-  </dropHandlers>
 </ArcGIS>
