+++ src/ArcGISProToArcGISEarth/SaveImageButton.cs
@@ -19,7 +19,7 @@ using System.Drawing;
 using System.IO;
 using System.Windows.Media.Imaging;
 
-namespace ToArcGISEarth
+namespace ArcGISProToArcGISEarth
 {
     public class SaveImageButton : Button
     {
