+++ src/ArcGISProToArcGISEarth/ToolHelper.cs
@@ -21,7 +21,7 @@ using System.Diagnostics;
 using System.IO;
 using System.Linq;
 
-namespace ToArcGISEarth
+namespace ArcGISProToArcGISEarth
 {
     public static class ToolHelper
     {
